refactor(core): extract compute→ledger callbacks to daemon composition root
Moves the three ledger-backed compute callback factories
(create_balance_callback, create_payment_callback,
create_commons_settlement_callback) and SettlementEngine construction
out of icn-core/supervisor/init_compute.rs into a new module:
bins/icnd/src/compute_wiring.rs.

Pre-built callbacks are injected into ComputeDeps via BootstrapHandles,
all as type-erased icn_compute callback types. icn-core never imports
icn_ledger. Adds NoOpSettlementQueryService to icn-kernel-api for
standalone/test fallback.

Meaning firewall ratchet: 21 → 5 (target met).
