+++ icn/bins/icnd/Cargo.toml
@@ -34,10 +34,13 @@ icn-trust.workspace = true
 icn-net.workspace = true
 icn-gossip.workspace = true
 icn-ledger.workspace = true
+icn-compute.workspace = true
 icn-ccl.workspace = true
 icn-store.workspace = true
 icn-rpc.workspace = true
 icn-obs.workspace = true
+hex = "0.4"
+serde_json.workspace = true
 
 # App crates (domain logic - lives outside kernel)
 icn-trust-app = { path = "../../../apps/trust" }
@@ -49,6 +52,9 @@ icn-charter-app = { path = "../../apps/charter" }
 # Needed for parameter store construction in daemon
 sled.workspace = true
 icn-governance.workspace = true
+
+[dev-dependencies]
+tempfile = "3.24"
 
 [lints]
 workspace = true
