+++ icn/crates/icn-kernel-api/src/lib.rs
@@ -112,12 +112,13 @@ pub use services::{
     FederationJoinResult, FederationService, FederationVouchRequest, FederationVouchResult,
     ForceCloseProposalRequest, ForceCloseProposalResult, FreezeMemberRequest, FreezeMemberResult,
     GovernanceEvent, GovernanceService, LedgerEvent, LedgerService, MembershipService,
-    RemoveMemberRequest, RemoveMemberResult, SecurityService, SecurityViolation, ServiceRegistry,
-    SettlementQueryResult, SettlementQueryService, SettlementReceiptResult, TreasuryEntryRequest,
-    TreasuryEntryResult, TreasuryOperationType as ServicesTreasuryOperationType, TrustClass,
-    TrustEvent, TrustService, UnfreezeMemberRequest, UnfreezeMemberResult, UpdateMemberRequest,
-    UpdateMemberResult, VetoProposalRequest, VetoProposalResult, TRUST_THRESHOLD_FEDERATED,
-    TRUST_THRESHOLD_KNOWN, TRUST_THRESHOLD_PARTNER,
+    NoOpSettlementQueryService, RemoveMemberRequest, RemoveMemberResult, SecurityService,
+    SecurityViolation, ServiceRegistry, SettlementQueryResult, SettlementQueryService,
+    SettlementReceiptResult, TreasuryEntryRequest, TreasuryEntryResult,
+    TreasuryOperationType as ServicesTreasuryOperationType, TrustClass, TrustEvent, TrustService,
+    UnfreezeMemberRequest, UnfreezeMemberResult, UpdateMemberRequest, UpdateMemberResult,
+    VetoProposalRequest, VetoProposalResult, TRUST_THRESHOLD_FEDERATED, TRUST_THRESHOLD_KNOWN,
+    TRUST_THRESHOLD_PARTNER,
 };
 pub use state::{
     BlobService, KvService, LogService, ObjectReplication, ReplicationPolicy, StateBackend,
